Reweighting functions that comprise an IRLS iteration

Reweighting functions that comprise an IRLS iteration

init_weight(object)

weight(object, data)

end_weight(object, tolerance = 1e-08)

Arguments

object
oomglm model.

data
an optional oomdata_tbl, oomdata_dbi, oomdata_con, tibble, data.frame, or list of observations to fit

tolerance
Tolerance used to determine convergence. Represents change in coefficient as a multiple of standard error.

Value

oomglm model

Details

The IRLS process comprises iterative calls to init_weight(), weight() and end_weight() which can be managed via a single call to fit().

These comprising functions are exposed for lower level control of the fitting process.

See also

Examples

# initiate `oomglm` model and `oomdata_tbl`
x <- oomglm(mpg ~ cyl + disp)
chunks <- oomdata_tbl(mtcars, chunk_size = 10)

# manually perform one round of IRLS and check results
x <- init_weight(x)
x <- weight(x, data = chunks)
x <- end_weight(x)

print(x$converged)

#> [1] FALSE

print(coef(x))

#> (Intercept)         cyl        disp 
#> 34.66099474 -1.58727681 -0.02058363 

# manually perform an additional round of IRLS
x <- init_weight(x)
x <- weight(x, data = chunks)
x <- end_weight(x)

print(x$converged)

#> [1] TRUE

print(coef(x))

#> (Intercept)         cyl        disp 
#> 34.66099474 -1.58727681 -0.02058363 

# the IRLS process as implemented by `fit()`
# is similar to the following
x <- oomglm(mpg ~ cyl + disp)
chunks <- oomdata_tbl(mtcars, chunk_size = 10)

while(!x$converged) {

  x <- init_weight(x)
  x <- weight(x, data = chunks)
  x <- end_weight(x)

}

tidy(x)

#> # A tibble: 3 x 7
#>   term        estimate std.error statistic  p.value conf.low  conf.high
#>   <chr>          <dbl>     <dbl>     <dbl>    <dbl>    <dbl>      <dbl>
#> 1 (Intercept)  34.7       2.55       13.6  4.02e-14  29.5     39.9     
#> 2 cyl          -1.59      0.712      -2.23 3.37e- 2  -3.04    -0.131   
#> 3 disp         -0.0206    0.0103     -2.01 5.42e- 2  -0.0416   0.000395