Iterate over tibble(), DBIResult or connection() in chunks.

Returns a function that repeats the following cycle: iteratively return chunk_size number of rows from data until data is exhausted; then return NULL once.

oomdata_tbl(data, chunk_size, ...)

oomdata_dbi(data, chunk_size, ...)

oomdata_con(data, chunk_size, header = TRUE, col_names = NULL, ...)

Arguments

data
tibble(), DBIResult, or connection() object

chunk_size
number of chunks to return with each iteration

...
ignored

header
logical, when TRUE colnames are determined from first row of connection. if FALSE col_names must be provided.

col_names
character vector. override col_names in first row of connection when headeris TRUE, or provide colnames if header is FALSE.

Details

oomdata_* functions are closures that return functions that iteratively return chunk_size number of rows from data until all rows have been returned. They will then return NULL once. They repeat this cycle ad-infinitum.

Examples

# `oomdata_tbl()` returns a function that when called will
# return `chunk_size` rows from a `tbl_df()`
chunks <- oomdata_tbl(mtcars, chunk_size = 16)

nrow(chunks())

#> [1] 16

nrow(chunks())

#> [1] 16

# when the data is exhausted the `oomdata_tbl()` function
# will return NULL once.
chunks()

#> NULL

# subsequent calls restart the cycle
nrow(chunks())

#> [1] 16

nrow(chunks())

#> [1] 16

chunks()

#> NULL

# `while` loops are useful for iterating over 
# `oomfeed` functions
while(!is.null(chunk <- chunks())){
  print(nrow(chunk))
}

#> [1] 16
#> [1] 16

# `oomdata_dbi()` returns a function that when called will
# return `chunk_size` rows from a query result set.
con <- DBI::dbConnect(RSQLite::SQLite(), path = ":dbname:")
dplyr::copy_to(con, mtcars, "mtcars", temporary = FALSE)
rs  <- DBI::dbSendQuery(con, "SELECT mpg, cyl, disp FROM mtcars")

chunks <- oomdata_dbi(rs, 16)

while(!is.null(chunk <- chunks())){
  print(nrow(chunk))
}

#> [1] 16
#> [1] 16

# ploom model functions automatically iterate over `oomdata` until 
# the source is exhausted (`oomlm()`, `oomlm_robust()`) or until 
# IRLS convergence (`oomglm()`, `oomglm_robust()`)
x <- oomlm(mpg ~ cyl + disp, chunks)

#> Error in init_oomlm(formula, weights): `weights` must be a formula

y <- iter_weight(oomglm(mpg ~ cyl + disp), chunks)

coef(x)

#> Error in coef(x): object 'x' not found

coef(y)

#> (Intercept)         cyl        disp 
#> 34.66099474 -1.58727681 -0.02058363