Initialize Updating Generalized Linear model

Perform generalized linear regression using Alan Miller's bounded memory QR factorization algorithm which enables models with p variables to be fit in p^2 memory.

oomglm(formula, data = NULL, family = gaussian(), weights = NULL,
  start = NULL, ...)

Arguments

formula
a symbolic description of the model to be fitted of class formula.

data
an optional oom_data, tibble, dataframe, list or environment.

family
A glm family object.

weights
a one-sided, single term formula specifying weights.

start
starting values for the parameters in the linear predictor.

...
ignored.

Value

It is up to the user to know when fitting is complete. Therefore, only basic model characteristics are provided as values with the oomglm object. Statistics are available on demand via summary and extractor functions.

converged

Indicates if the IRLS algorithm has converged.

iter

The number of iterations of IRLS performed.

n

The number observations processed per round of IRLS.

df.residual

The residual degrees of freedom.

df.null

The residual degrees of freedom.

formula

the formula() object specifying the linear model.

family

a family() object describing the error distribution and link function used in the model.

terms

The terms() object used.

weights

The weights formula() provided to the model.

call

The matched call.

Details

oomglm initializes an object of class oomglm inheriting from the class oomlm. oomglm objects are intended to be iteratively updated with new data via calls to update(). Iterative fitting over all data updates are performed with the function iter_weight(). If data is provided to the oomglm() function call, an update() round will be performed on initialization.

A oomglm object can be in various states of fit depending on the number of seen observations and rounds of IRLS that have been performed. It is important to view the model within the context of: the number of observations processed per round of IRLS (n); the number of IRLS iterations that have been performed (iter); and if the IRLS algorithm has converged (converged).

See also

Examples

# The `oomglm()` function fits generalized linear models via 
# Iteratively Weighted Least Squares (IWLS).  

# When fitting in-memory data the process is similar to `oomlm()` but we use the 
# function `iter_weight()` instead of `update()`. `iter_weight()` fits the model
# via iterative passes over the data until convergence.

# initialize the model
x <- oomglm(mpg ~ cyl + disp)

# re-weight 8 times or until convergence
x <- iter_weight(x, mtcars, max_iter = 8)

# To fit data in chunks, use `oom_data()`:

# initialize the model
x    <- oomglm(mpg ~ cyl + disp)
feed <- oom_data(mtcars, chunk_size = 10)

# iteratively reweight model
x <- iter_weight(x, feed, max_iter = 8)

# The `iter_weight()` process can also be implemented directly with
# component `ploom` functions:

x    <- oomglm(mpg ~ cyl + disp)
feed <- oom_data(mtcars, chunk_size = 10)

# a first pass over the data
x <- init_weight(x)
x <- weight(x, feed)
x <- end_weight(x)
x

#> 
#> Call:  oomglm(mpg ~ cyl + disp)
#> 
#> Coefficients:
#> (Intercept)          cyl         disp  
#>    34.66099     -1.58728     -0.02058  
#> 
#> Observations included:  32 
#> Degrees of Freedom: 31 Total (i.e. Null);  29 Residual
#> Residual Deviance: 270.7 	AIC: 167.1 
#> 
#> Converged: FALSE 
#> Number of Fisher Scoring iterations: 1 

# a second pass over the data
x <- init_weight(x)
x <- weight(x, feed)
x <- end_weight(x)
x

#> 
#> Call:  oomglm(mpg ~ cyl + disp)
#> 
#> Coefficients:
#> (Intercept)          cyl         disp  
#>    34.66099     -1.58728     -0.02058  
#> 
#> Observations included:  32 
#> Degrees of Freedom: 31 Total (i.e. Null);  29 Residual
#> Residual Deviance: 270.7 	AIC: 167.1 
#> 
#> Converged: TRUE 
#> Number of Fisher Scoring iterations: 2 

# This is meant to be useful when debugging / evaluating models with long 
# runtimes by exposing the individual steps of the model process for inspection.