Out of Memory Generalized Linear model with robust standard errors

Perform linear regression via Alan Miller's bounded memory QR factorization algorithm enables models with p variables with p variables to be fit with robust standard errors in p^4 memory.

oomglm_robust(formula, data = NULL, family = gaussian(),
  weights = NULL, start = NULL, se_type = "H1", ...)

Arguments

formula
a symbolic description of the model to be fitted of class formula.

data
an optional oom_data, tibble, data.frame, list or environment.

family
A glm family object.

weights
a one-sided, single term formula specifying weights.

start
starting values for the parameters in the linear predictor.

se_type
string indicating what se type to usecan be "HC0", "HC1" "stata", or "classical". default "HC1".

...
ignored.

Value

It is up to the user to know when fitting is complete. Therefore, only basic model characteristics are provided as values with the oomglm object. Statistics are available on demand via summary and extractor functions.

converged

Indicates if the IRLS algorithm has converged.

iter

The number of iterations of IRLS performed.

n

The number observations processed per round of IRLS.

df.residual

The residual degrees of freedom.

df.null

The residual degrees of freedom.

formula

the formula() object specifying the linear model.

family

a family() object describing the error distribution and link function used in the model.

terms

The terms() object used.

weights

The weights formula() provided to the model.

call

The matched call.

Details

oomglm_robust initializes an object of class oomlm. oomlm objects are intended to be iteratively updated with new data via the function update(). If data is provided to the oolm function call, an update() round will be performed on initialization.

The provided formula must not contain any data-dependent terms to ensure consistency across calls to update(). Factors are permitted, but the levels of the factor must be the same across all data chunks. Empty factor levels are accepted.

See also