Out of memory Linear model

Perform linear regression via Alan Miller's bounded memory QR factorization algorithm which enables models with p variables to be fit in p^2 memory.

oomlm(formula, data = NULL, weights = NULL, ...)

Arguments

formula
a symbolic description of the model to be fitted of class formula.

data
an optional oom_data, tibble, data.frame, list or environment.

weights
a one-sided, single term formula specifying weights.

...
ignored.

Value

A oomlm object is perpetually in an in-progress state. It is up to the user to know when fitting is complete. Therefore, only basic model characteristics are provided as values with the oolm object. Statistics are available on demand via summary and extractor functions.

n

the number of observations processed.

df.residual

the residual degrees of freedom.

formula

the stats::formula() object specifying the linear model.

terms

the stats::terms() object specifying the terms of the linear model.

weights

a one-sided, single term formula specifying weights.

call

the matched call.

Details

oomlm initializes an object of class oomlm. oomlm objects are intended to be iteratively updated with new data via the function update(). If data is provided to the oolm function call, an update() round will be performed on initialization.

The provided formula must not contain any data-dependent terms to ensure consistency across calls to update(). Factors are permitted, but the levels of the factor must be the same across all data chunks. Empty factor levels are accepted.

See also

Examples

# The `ploom` linear model, `oomlm`, is similar to base `lm` for fitting
# in-memory data.

x <- oomlm(mpg ~ cyl + disp, data = mtcars)

# Models are initalized with a call to `oomlm()` and updated with `update()`.
# The intended pattern is to initialize models without referencing data,
# then call `update()` on each data chunk.

# proxy for big data feed 
chunks  <- oom_data(mtcars, chunk_size = 10)

# initialize the model
x <- oomlm(mpg ~ cyl + disp)

# iteratively update model with data chunks
while(!is.null(chunk <- chunks())) {
  x <- update(x, chunk)
}

summary(x)

#> 
#> Call:  oomlm(mpg ~ cyl + disp)
#> 
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept) 34.66099    2.54700  13.609 4.02e-14 ***
#> cyl         -1.58728    0.71184  -2.230   0.0337 *  
#> disp        -0.02058    0.01026  -2.007   0.0542 .  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Observations included:  32 
#> Residual standard error: 3.055 on 29 degrees of freedom
#> Multiple R-squared:  0.7596,	Adjusted R-squared:  0.743 
#> F-statistic: 45.81 on 2 and 29 DF,  p-value: 1.058e-09
#>