Out of memory Generalized Linear model

Perform generalized linear regression using Alan Miller's bounded memory QR factorization algorithm which enables models with p variables to be fit in p^2 memory.

oomglm(formula, family = gaussian(), weights = NULL, start = NULL,
  ...)

Arguments

formula
a symbolic description of the model to be fitted of class formula().

family
a family() object.

weights
a one-sided, single term formula() specifying weights.

start
starting values for the parameters in the linear predictor.

...
ignored.

data
an optional oomdata_tbl(), oomdata_dbi(), oomdata_con(), tibble(), data.frame(), or list() of observations to fit

Value

It is up to the user to know when fitting is complete. Therefore, only basic model characteristics are provided as values. Statistics are available on demand via summary and extractor functions.

converged

Indicates if the IRLS algorithm has converged.

iter

The number of iterations of IRLS performed.

n

The number observations processed per round of IRLS.

df.residual

The residual degrees of freedom.

df.null

The residual degrees of freedom.

formula

the formula() object specifying the linear model.

family

a family() object describing the error distribution and link function used in the model.

terms

The terms() object used.

weights

The weights formula() provided to the model.

call

The matched call.

Details

An oomglm() model can be in various states of fit depending on the number of seen observations and rounds of IRLS that have been performed. It is important to view the model within the context of: the number of observations processed per round of IRLS (n); the number of IRLS iterations that have been performed (iter); and if the IRLS algorithm has converged (converged).

See also

Examples

# The `oomglm()` function employs Iteratively Weighted Least Squares (IWLS).
# The IWLS iterations are performed by the function `iter_weight()` which
# makes passes over the data until estimate convergence.

# reweight 4 times or until convergence
x <- oomglm(mpg ~ cyl + disp)
x <- iter_weight(x, mtcars, max_iter = 4)

tidy(x)

#> # A tibble: 3 x 7
#>   term        estimate std.error statistic  p.value conf.low  conf.high
#>   <chr>          <dbl>     <dbl>     <dbl>    <dbl>    <dbl>      <dbl>
#> 1 (Intercept)  34.7       2.55       13.6  4.02e-14  29.5     39.9     
#> 2 cyl          -1.59      0.712      -2.23 3.37e- 2  -3.04    -0.131   
#> 3 disp         -0.0206    0.0103     -2.01 5.42e- 2  -0.0416   0.000395

# To fit data in chunks, use `oomdata_tbl()`:

y      <- oomglm(mpg ~ cyl + disp)
chunks <- oomdata_tbl(mtcars, chunk_size = 10)
y      <- iter_weight(y, chunks, max_iter = 4)

tidy(y)

#> # A tibble: 3 x 7
#>   term        estimate std.error statistic  p.value conf.low  conf.high
#>   <chr>          <dbl>     <dbl>     <dbl>    <dbl>    <dbl>      <dbl>
#> 1 (Intercept)  34.7       2.55       13.6  4.02e-14  29.5     39.9     
#> 2 cyl          -1.59      0.712      -2.23 3.37e- 2  -3.04    -0.131   
#> 3 disp         -0.0206    0.0103     -2.01 5.42e- 2  -0.0416   0.000395