Out of memory Linear model

Perform memory-efficient linear regression using the AS274 bounded memory QR factorization algorithm.

oomlm(formula, weights = NULL, ...)

Arguments

formula
A symbolic description of the model to be fitted of class formula.

weights
A one-sided, single term formula specifying weights.

...
Ignored.

Value

An oomlm model is perpetually in an in-progress state. It is up to the user to know when fitting is complete. Therefore, only basic model characteristics are provided as values. Statistics are available on demand via summary and extractor functions.

n

The number of observations processed.

df.residual

The residual degrees of freedom.

formula

The formula object specifying the linear model.

terms

The terms object specifying the terms of the linear model.

weights

A one-sided, single term formula specifying weights.

call

The matched call.

Details

The provided formula must not contain any data-dependent terms to ensure consistency across calls to fit(). Factors are permitted, but the levels of the factor must be the same across all data chunks. Empty factor levels are accepted.

See also

Examples

# `oomlm` are defined with a call to `oomlm()` and fit to data
# with a call to `fit()`
x <- oomlm(mpg ~ cyl + disp)
x <- fit(x, mtcars)
print(x)

#> 
#> Call:  oomlm(mpg ~ cyl + disp)
#> 
#> Coefficients:
#> (Intercept)          cyl         disp  
#>    34.66099     -1.58728     -0.02058  
#> 
#> Observations included:  32 

# `oomlm` models can be fit with more data via subsequent calls
# to the `fit()` function
chunks <- purrr::pmap(mtcars, list)

y <- oomlm(mpg ~ cyl + disp)

for(chunk in chunks) {
  y <- fit(y, chunk)
}

tidy(x)

#> # A tibble: 3 x 7
#>   term        estimate std.error statistic  p.value conf.low  conf.high
#>   <chr>          <dbl>     <dbl>     <dbl>    <dbl>    <dbl>      <dbl>
#> 1 (Intercept)  34.7       2.55       13.6  4.02e-14  29.5     39.9     
#> 2 cyl          -1.59      0.712      -2.23 3.37e- 2  -3.04    -0.131   
#> 3 disp         -0.0206    0.0103     -2.01 5.42e- 2  -0.0416   0.000395

# `oomdata_tbl()` facilitates iterating through data rows in chunks
chunks  <- oomdata_tbl(mtcars, chunk_size = 1)

# `fit()` will automatically fit over all chunks in an `oomdata`
# object
z <- oomlm(mpg ~ cyl + disp)
z <- fit(z, data = chunks)
summary(z)

#> 
#> Call:  oomlm(mpg ~ cyl + disp)
#> 
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept) 34.66099    2.54700  13.609 4.02e-14 ***
#> cyl         -1.58728    0.71184  -2.230   0.0337 *  
#> disp        -0.02058    0.01026  -2.007   0.0542 .  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Observations included:  32 
#> Residual standard error: 3.055 on 29 degrees of freedom
#> Multiple R-squared:  0.7596,	Adjusted R-squared:  0.743 
#> F-statistic: 45.81 on 2 and 29 DF,  p-value: 1.058e-09
#>